backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
httpx>=0.28.0
pydantic>=2.0.0,<3.0.0
certifi
sqlalchemy>=2.0.0
python-dotenv
